java/src/main/java/dms/pastor/examples/asynchronous/CompletableFutureExample.java
package dms.pastor.examples.asynchronous;

import dms.pastor.domain.exception.SomethingWentWrongException;

import java.util.concurrent.CompletableFuture;
import java.util.concurrent.ExecutionException;
import java.util.concurrent.Executors;
import java.util.concurrent.Future;
import java.util.stream.Collectors;
import java.util.stream.Stream;

/**
 * CompletableFuture is used for to perform an asynchronous computation
 */
public class CompletableFutureExample {

    public Future<String> calculateAsync() {
        CompletableFuture<String> completableFuture = new CompletableFuture<>();

        Executors.newCachedThreadPool().submit(() -> {
            Thread.sleep(500);
            completableFuture.complete("Hello");
            return null;
        });

        return completableFuture;
    }

    public String parallelRunningOfMultipleFutures() throws ExecutionException, InterruptedException {
        CompletableFuture<String> future1
                = CompletableFuture.supplyAsync(() -> "Hello");
        CompletableFuture<String> future2
                = CompletableFuture.supplyAsync(() -> "Beautiful");
        CompletableFuture<String> future3
                = CompletableFuture.supplyAsync(() -> "World");

        CompletableFuture<Void> combinedFuture
                = CompletableFuture.allOf(future1, future2, future3);

        combinedFuture.get();

        return Stream.of(future1, future2, future3)
                .map(CompletableFuture::join)
                .collect(Collectors.joining(" "));

    }

    public CompletableFuture<String> errorHandlingDuringInCompletableFuture(String name) {
        return CompletableFuture.supplyAsync(() -> {
            if (name == null) {
                throw new SomethingWentWrongException("Whoops");
            }
            return "Hello, " + name;
        }).handle((s, t) -> s != null ? s : "Hello, Stranger!");


    }


}
